[TNT-245] move: 패키지 분리
